add basic frontend
