+++ modules/notepad.html
@@ -1,0 +1,3 @@
+<div class="list">
+	<div class="list-item">You are in notepad</div>
+</div>
