+++ public/includes/frames/error.ejs
@@ -1,0 +1,5 @@
+<div class="list">
+		<div class="list-item">
+			<%- message || 'An error occured' %>
+		</div>
+</div>
