+++ public/includes/frames/taskbar.ejs
@@ -1,0 +1,24 @@
+<div class="taskbar-item">
+	<a href="/">
+		<img src="assets/icons/home.png">
+		<span>Home</span>
+	</a>
+</div>
+<div class="taskbar-item">
+	<a href="/new?q=calendar">
+		<img src="assets/icons/unknown.png">
+		<span>Calendar</span>
+	</a>
+</div>
+<div class="taskbar-item">
+	<a href="/new?q=clock">
+		<img src="assets/icons/unknown.png">
+		<span>Clock</span>
+	</a>
+</div>
+<div class="taskbar-item">
+	<a href="/new?q=notepad">
+		<img src="assets/icons/unknown.png">
+		<span>Notepad</span>
+	</a>
+</div>
