+++ public/includes/templates/window.ejs
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+	<meta charset="utf-8">
+	<meta name="viewport" content="width=device-width">
+	<title>Gigadex<%= name || '' %></title>
+	<link href="style.css" rel="stylesheet" type="text/css" />
+	<link rel="shortcut icon" href="assets/images/fav.png" type="image/png">
+</head>
+
+<body>
+	<div class="app">
+		<div class="window">
+			<%- content %>
+		</div>
+		<div class="taskbar">
+			<%- include('../frames/taskbar') %>
+		</div>
+	</div>
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+	<script src="script.js"></script>
+</body>
+
+</html>
