+++ public/views/index.ejs
@@ -4,7 +4,7 @@
 <head>
 	<meta charset="utf-8">
 	<meta name="viewport" content="width=device-width">
-	<title>Title</title>
+	<title>Gigadex</title>
 	<link href="style.css" rel="stylesheet" type="text/css" />
 </head>
