+++ public/views/index.html
@@ -1,0 +1,4 @@
+<div class="list">
+	<div class="list-item">No apps open</div>
+	<div class="list-item"><a href="/new">Open an app</a></div>
+</div>
