+++ public/views/login.html
@@ -1,0 +1,7 @@
+<div class="login">
+	<form>
+		<input type="text" name="username" placeholder="Username" autocomplete="off">
+		<input type="password" name="password" placeholder="Password">
+		<input type="submit" value="Login">
+	</form>
+</div>
