+++ public/views/new.ejs
@@ -1,0 +1,5 @@
+<div class="list">
+	<% for (let i in apps) { %>
+		<div class="list-item"><a href="/new?q=<%= apps[i] %>"><%= apps[i] %></a></div>
+	<% } %>
+</div>
